+++ tracing/gemini/src/jvmMain/kotlin/org/jetbrains/ai/tracy/gemini/adapters/GeminiLLMTracingAdapter.kt
@@ -59,8 +59,10 @@ class GeminiLLMTracingAdapter : LLMTracingAdapter(genAISystem = GenAiSystemIncub
 
     override fun getSpanName(request: TracyHttpRequest) = "Gemini-generation"
 
-    // streaming is not supported
-    override fun isStreamingRequest(request: TracyHttpRequest) = false
+    override fun isStreamingRequest(request: TracyHttpRequest): Boolean {
+        val (_, operation) = request.url.modelAndOperation()
+        return operation == "streamGenerateContent"
+    }
     override fun handleStreaming(span: Span, url: TracyHttpUrl, events: String) {
         val handler = selectHandler(url)
         handler.handleStreaming(span, events)